# packages/api-types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sahaayak",
    "description": "Local-language voice agent for discovering government schemes, scholarships, and jobs, and checking eligibility for them.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/benefits/{benefit_id}": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Benefit Detail",
        "description": "Return one active benefit and its source/review metadata.",
        "operationId": "benefit_detail_api_benefits__benefit_id__get",
        "parameters": [
          {
            "name": "benefit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Benefit Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenefitDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/escalations/{ticket_id}/route": {
      "post": {
        "tags": [
          "escalations"
        ],
        "summary": "Route Ticket",
        "operationId": "route_ticket_api_escalations__ticket_id__route_post",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticket Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/departments": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List Directory Entries",
        "operationId": "list_directory_entries_api_admin_departments_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "pending",
                "approved",
                "rejected",
                "all"
              ],
              "type": "string",
              "default": "all",
              "title": "Status"
            }
          },
          {
            "name": "state_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State Code"
            }
          },
          {
            "name": "district_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "District Name"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 200,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectoryListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/departments/{entry_id}/approve": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Approve Directory Entry",
        "operationId": "approve_directory_entry_api_admin_departments__entry_id__approve_post",
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entry Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DirectoryDecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectoryEntryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/departments/{entry_id}/deactivate": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Deactivate Directory Entry",
        "operationId": "deactivate_directory_entry_api_admin_departments__entry_id__deactivate_post",
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entry Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DirectoryDecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectoryEntryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BenefitDetailOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "domain": {
            "$ref": "#/components/schemas/Domain"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "state_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State Code"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "benefits_text": {
            "type": "string",
            "title": "Benefits Text"
          },
          "documents_required": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Documents Required"
          },
          "application_process": {
            "type": "string",
            "title": "Application Process"
          },
          "eligibility_initial": {
            "additionalProperties": true,
            "type": "object",
            "title": "Eligibility Initial"
          },
          "verification_status": {
            "$ref": "#/components/schemas/VerificationStatus"
          },
          "source_title": {
            "type": "string",
            "title": "Source Title"
          },
          "source_document_url": {
            "type": "string",
            "title": "Source Document Url"
          },
          "source_excerpt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Excerpt"
          },
          "verified_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified At"
          },
          "last_verified_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Verified Date"
          },
          "valid_from": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid From"
          },
          "valid_until": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid Until"
          },
          "job_metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Job Metadata",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "id",
          "domain",
          "name",
          "state_code",
          "category",
          "description",
          "benefits_text",
          "documents_required",
          "application_process",
          "eligibility_initial",
          "verification_status",
          "source_title",
          "source_document_url",
          "source_excerpt",
          "verified_at",
          "last_verified_date",
          "valid_from",
          "valid_until"
        ],
        "title": "BenefitDetailOut",
        "description": "Public benefit details with enough provenance to make trust inspectable."
      },
...
      "DirectoryDecisionRequest": {
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 3,
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "DirectoryDecisionRequest"
      },
      "DirectoryEntryOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "entry_key": {
            "type": "string",
            "title": "Entry Key"
          },
          "state_code": {
            "type": "string",
            "title": "State Code"
          },
          "district_code": {
            "type": "string",
            "title": "District Code"
          },
          "district_name": {
            "type": "string",
            "title": "District Name"
          },
          "service_domain": {
            "type": "string",
            "title": "Service Domain"
          },
          "pincode": {
            "type": "string",
            "title": "Pincode"
          },
          "pincode_prefix": {
            "type": "string",
            "title": "Pincode Prefix"
          },
          "department_code": {
            "type": "string",
            "title": "Department Code"
          },
          "department_name": {
            "type": "string",
            "title": "Department Name"
          },
          "help_centre_name": {
            "type": "string",
            "title": "Help Centre Name"
          },
          "address": {
            "type": "string",
            "title": "Address"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "website_url": {
            "type": "string",
            "title": "Website Url"
          },
          "source_name": {
            "type": "string",
            "title": "Source Name"
          },
          "source_url": {
            "type": "string",
            "title": "Source Url"
          },
          "source_record_id": {
            "type": "string",
            "title": "Source Record Id"
          },
          "source_last_verified": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Last Verified"
          },
          "approval_status": {
            "type": "string",
            "title": "Approval Status"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "stale": {
            "type": "boolean",
            "title": "Stale"
          },
          "priority": {
            "type": "integer",
            "title": "Priority"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "entry_key",
          "state_code",
          "district_code",
          "district_name",
          "service_domain",
          "pincode",
          "pincode_prefix",
          "department_code",
          "department_name",
          "help_centre_name",
          "address",
          "phone",
          "email",
          "website_url",
          "source_name",
          "source_url",
          "source_record_id",
          "source_last_verified",
          "approval_status",
          "is_active",
          "stale",
          "priority",
          "created_at",
          "updated_at"
        ],
        "title": "DirectoryEntryOut"
      },
      "DirectoryListOut": {
        "properties": {
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "stale_after_days": {
            "type": "integer",
            "title": "Stale After Days"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/DirectoryEntryOut"
            },
            "type": "array",
            "title": "Entries"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "status_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Status Counts"
          }
        },
        "type": "object",
        "required": [
          "generated_at",
          "stale_after_days",
          "entries",
          "total",
          "status_counts"
        ],
        "title": "DirectoryListOut"
      },
      "Domain": {
        "type": "string",
        "enum": [
          "scheme",
          "scholarship",
          "job"
        ],
        "title": "Domain"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RouteRequest": {
        "properties": {
          "department": {
            "type": "string",
            "maxLength": 160,
            "minLength": 2,
            "title": "Department"
          },
          "routing_location": {
            "type": "string",
            "maxLength": 120,
            "title": "Routing Location",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "department"
        ],
        "title": "RouteRequest"
      },
...
      "TicketOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "caller_context": {
            "additionalProperties": true,
            "type": "object",
            "title": "Caller Context"
          },
          "transcript_excerpt": {
            "type": "string",
            "title": "Transcript Excerpt"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "assigned_to": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned To"
          },
          "claimed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Claimed At"
          },
          "sla_due_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sla Due At"
          },
          "sla_breached": {
            "type": "boolean",
            "title": "Sla Breached",
            "default": false
          },
          "department": {
            "type": "string",
            "title": "Department"
          },
          "routing_location": {
            "type": "string",
            "title": "Routing Location"
          },
          "routing_source": {
            "type": "string",
            "title": "Routing Source"
          },
          "routing_directory_entry_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Routing Directory Entry Id"
          },
          "routing_source_url": {
            "type": "string",
            "title": "Routing Source Url",
            "default": ""
          },
          "routing_verified_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Routing Verified At"
          },
          "operator_notes": {
            "items": {
              "$ref": "#/components/schemas/TicketNoteOut"
            },
            "type": "array",
            "title": "Operator Notes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "resolved_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "resolved_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved By"
          },
          "resolution_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Code"
          },
          "resolution_note": {
            "type": "string",
            "title": "Resolution Note",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "id",
          "session_id",
          "reason",
          "transcript_excerpt",
          "status",
          "department",
          "routing_location",
          "routing_source",
          "created_at"
        ],
        "title": "TicketOut"
      },
...
    }
  }
}